backend/requirement.txt
fastapi>=0.110.0,<1.0.0
uvicorn[standard]>=0.27.0,<1.0.0
pydantic>=2.6.0,<3.0.0
pydantic-settings>=2.2.0,<3.0.0
email-validator>=2.1.0,<3.0.0
PyJWT>=2.8.0,<3.0.0
slowapi>=0.1.9,<1.0.0
numpy>=1.26.0,<3.0.0
scikit-learn>=1.4.0,<2.0.0
httpx>=0.27.0,<1.0.0
motor>=3.6.0,<4.0.0
passlib[bcrypt]>=1.7.4,<2.0.0
redis>=5.0.0,<6.0.0
openai>=1.40.0,<2.0.0
google-auth>=2.29.0,<3.0.0
reportlab>=4.2.0,<5.0.0
aiosmtplib>=3.0.0,<4.0.0
python-multipart>=0.0.9,<1.0.0
pytest>=8.0.0,<9.0.0
pytest-asyncio>=0.23.0,<1.0.0
google-generativeai>=0.4.0
PyPDF2>=3.0.0
